update DicomJSONDataSource to support directURL method
